+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,8 +6,6 @@ import java.util.Set;
 import java.util.TreeSet;
 import java.util.stream.Collectors;
 
-import seedu.address.logic.parser.ParserUtil;
-import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.ReadOnlyRecipeBook;
 import seedu.address.model.RecipeBook;
 import seedu.address.model.goal.Goal;
@@ -17,12 +15,9 @@ import seedu.address.model.recipe.Step;
 import seedu.address.model.recipe.Time;
 
 import seedu.address.model.recipe.ingredient.Grain;
-import seedu.address.model.recipe.ingredient.Ingredient;
 import seedu.address.model.recipe.ingredient.Other;
 import seedu.address.model.recipe.ingredient.Protein;
 import seedu.address.model.recipe.ingredient.Quantity;
-import seedu.address.model.recipe.ingredient.QuantityUtil;
-import seedu.address.model.recipe.ingredient.Unit;
 import seedu.address.model.recipe.ingredient.Vegetable;
 
 /**
