+++ src/main/java/seedu/address/storage/JsonAdaptedGrain.java
@@ -4,12 +4,9 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.logic.parser.ParserUtil;
-import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.recipe.ingredient.Grain;
 import seedu.address.model.recipe.ingredient.Quantity;
-import seedu.address.model.recipe.ingredient.QuantityUtil;
-import seedu.address.model.recipe.ingredient.Unit;
+import seedu.address.model.util.QuantityUtil;
 
 /**
  * Jackson-friendly version of {@link Grain}.
