+++ src/main/java/seedu/address/storage/JsonAdaptedOther.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonValue;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.recipe.ingredient.Other;
 import seedu.address.model.recipe.ingredient.Quantity;
-import seedu.address.model.recipe.ingredient.QuantityUtil;
+import seedu.address.model.util.QuantityUtil;
 
 /**
  * Jackson-friendly version of {@link Other}.
